TextValue: Moved to a separate module
Also update changelog and readme
