+++ resources-ktx/build.gradle.kts
@@ -5,13 +5,16 @@ plugins {
     id("com.redmadrobot.publish")
 }
 
+version = "1.3.1-0"
+description = "A set of Kotlin extensions for accessing resources"
+
+android {
+    namespace = "com.redmadrobot.extensions.resources-ktx"
+}
+
 dependencies {
-    api(projects.resources.resourcesCommon)
+    api(projects.textvalue)
     api(androidx.fragment)
     implementation(androidx.core)
     implementation(androidx.appcompat.resources)
-}
-
-android {
-    namespace = "com.redmadrobot.extensions.resources"
 }
