+++ resources-ktx/src/main/kotlin/Attributes.kt
