--- resources/build.gradle.kts
@@ -1,4 +1,0 @@
-subprojects {
-    version = "1.3.1-0"
-    description = "A set of Kotlin extensions for accessing resources"
-}
