+++ textvalue/CHANGELOG.md
@@ -1,0 +1,5 @@
+## Unreleased
+
+### Changed
+
+- Create separate modules textvalue and textvalue-compose
